add v2 api paths
