+++ src/config/config.go
@@ -22,8 +22,8 @@ var (
 	OauthEnable, _    = strconv.ParseBool(getEnv("ADMINIO_OAUTH_ENABLE", "false"))
 	AuditLogEnable, _ = strconv.ParseBool(getEnv("ADMINIO_AUDIT_LOG_ENABLE", "false"))
 	OauthProvider     = getEnv("ADMINIO_OAUTH_PROVIDER", "github")
-	OauthClientId     = getEnv("ADMINIO_OAUTH_CLIENT_ID", "111")
-	OauthClientSecret = getEnv("ADMINIO_OAUTH_CLIENT_SECRET", "222")
+	OauthClientId     = getEnv("ADMINIO_OAUTH_CLIENT_ID", "my-github-oauth-app-client-id")
+	OauthClientSecret = getEnv("ADMINIO_OAUTH_CLIENT_SECRET", "my-github-oauth-app-secret")
 	OauthCallback     = getEnv("ADMINIO_OAUTH_CALLBACK", "http://"+ServerHostPort+"/auth/callback")
 	OauthCustomDomain = getEnv("ADMINIO_OAUTH_CUSTOM_DOMAIN", "")
 )
