+++ src/go.mod
@@ -17,6 +17,7 @@ require (
 	github.com/moul/http2curl v1.0.0 // indirect
 	github.com/onsi/ginkgo v1.12.0 // indirect
 	github.com/onsi/gomega v1.9.0 // indirect
+	github.com/rzrbld/goth-provider-wso2 v0.0.0-20200321083654-32bbe73a67d4
 	github.com/sergi/go-diff v1.1.0 // indirect
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
 	github.com/valyala/fasthttp v1.9.0 // indirect
